src/requirements.txt
Flask==2.3.3
flasgger>=0.9.6
Pillow==10.0.0